tuan01-02/management/serializers/device.py
from rest_framework import serializers
from management.models import Product

class ProductSerializer(serializers.ModelSerializer):
    class Meta:
        model = Product
        fields = '__all__'

