tuan01-02/requirements.txt
django==4.1
gunicorn==20.1.0
djangorestframework
django-cors-headers==3.13.0
djangorestframework-simplejwt==5.2.0
mysqlclient
